Added multiple maps support + code style
I tried a "all base maps any%" run, but the autosplitter didn't work, because it did all splits when the first map was completed.
As quick solution the player now has a "safety minute" to leave the finish screen. 

I added "current.finished" as isLoading condition because the timer was already running for the next split (map) at the finish screen.